requirements (1).txt

fastapi
uvicorn
pytesseract
python-multipart
pillow
fpdf
python-dotenv
